@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.15.1 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -78,6 +78,7 @@ flutter:
   assets:
     - assets/images/
     - assets/icons/
+    - assets/google_fonts/
 
   # To add assets to your application, add an assets section, like this:
   # assets:
