@@ -1,3 +1,7 @@
-const baseUrl = "http://192.168.1.8:8080";
-const socketUrl ="ws://192.168.1.8:8080/tc-socket";
+const baseUrl = "http://192.168.1.5:8080";
+
+const kSocketUrl ="ws://192.168.1.5:8088/tc-socket";
+
 const imageUrl = baseUrl + "/api/v1/public/images/";
+
+
