@@ -3,8 +3,6 @@ import 'package:flutter/material.dart';
 import '../../constants/app_theme_constants.dart';
 
 class ChatController extends StatelessWidget {
-
-
   const ChatController({
     Key? key,
     required this.size,
@@ -12,6 +10,7 @@ class ChatController extends StatelessWidget {
     required this.onSendBtnPressed,
     this.focusNode,
     this.onChange,
+    required this.isTextMessageEmpty,
   }) : super(key: key);
 
   final Size size;
@@ -19,14 +18,15 @@ class ChatController extends StatelessWidget {
   final FocusNode? focusNode;
   final Function onSendBtnPressed;
   final Function? onChange;
+  final bool isTextMessageEmpty;
 
   @override
   Widget build(BuildContext context) {
     return Container(
-      // height: 56,
-      constraints: const BoxConstraints(
-        minHeight: 56,
-      ),
+      height: kChatControllerHeight,
+      // constraints: const BoxConstraints(
+      //   minHeight: kChatControllerHeight,
+      // ),
       decoration: const BoxDecoration(
         color: kPrimaryColor,
       ),
@@ -34,21 +34,44 @@ class ChatController extends StatelessWidget {
         crossAxisAlignment: CrossAxisAlignment.center,
         mainAxisAlignment: MainAxisAlignment.center,
         children: [
+          IconButton(
+            alignment: Alignment.center,
+            constraints: const BoxConstraints(),
+            padding: const EdgeInsets.symmetric(horizontal: 10),
+            onPressed: () {},
+            icon: const Icon(
+              Icons.face_retouching_natural,
+              color: Colors.white,
+            ),
+          ),
           Expanded(
             child: Container(
               alignment: Alignment.center,
-              width: size.width * .8,
-              padding: const EdgeInsets.symmetric(horizontal: 10),
+              // constraints: BoxConstraints(
+              //   minHeight: kChatControllerHeight,
+              // ),
+              height: kChatControllerHeight - 10,
+              decoration: BoxDecoration(
+                  color: Colors.grey.shade50,
+                  borderRadius: BorderRadius.circular(40)),
+              // margin: const EdgeInsets.symmetric(horizontal: 5),
               child: TextField(
-                style: const TextStyle(color: Colors.white),
+                style: const TextStyle(color: kPrimaryColor),
+                keyboardType: TextInputType.multiline,
+                textAlign: TextAlign.left,
+                // minLines: null,
+                maxLines: null,
+                textInputAction: TextInputAction.next,
+                // expands: true,
+
                 onChanged: (string) =>
                     onChange != null ? onChange!(string) : () {},
                 focusNode: focusNode,
                 controller: messageController,
                 decoration: InputDecoration(
-                  isDense: true,
-                  contentPadding: const EdgeInsets.symmetric(vertical: 8),
-                  // border: InputBorder.none,
+                  // isDense: true,
+                  contentPadding: const EdgeInsets.symmetric(horizontal: 8),
+                  border: InputBorder.none,
                   hintStyle: TextStyle(
                     color: Colors.white.withOpacity(.55),
                   ),
@@ -59,18 +82,12 @@ class ChatController extends StatelessWidget {
           ),
           IconButton(
             alignment: Alignment.center,
-            onPressed: () {},
-            icon: const Icon(
-              Icons.face_retouching_natural,
-              color: Colors.white,
-            ),
-          ),
-          IconButton(
-            alignment: Alignment.center,
-            onPressed: () => onSendBtnPressed(),
-            icon: const Icon(
+            constraints: const BoxConstraints(),
+            padding: const EdgeInsets.symmetric(horizontal: 10),
+            onPressed: () => !isTextMessageEmpty ? onSendBtnPressed() : null,
+            icon: Icon(
               Icons.send,
-              color: Colors.white,
+              color: isTextMessageEmpty ? Colors.grey : Colors.white,
             ),
           ),
         ],
